feat: validate web server configuration at startup
Split config parsing (configSchema.ts) from the parsed singleton
(config.ts), and add a Nitro startup plugin that parses the
environment before the port binds. A misconfigured instance now
exits non-zero with every offending key and reason logged, rather
than starting and failing each request behind a healthy-looking
listener.
